Moved classes around; updated documentation of anything but the reduction classes.
