+++ WPFSKillTree/TreeGenerator/Algorithm/Model/TwoDArray.cs
@@ -1,0 +1,49 @@
+﻿using System;
+
+namespace POESKillTree.TreeGenerator.Algorithm.Model
+{
+    /// <summary>
+    /// Read only representation of a two-dimensional array of type T.
+    /// </summary>
+    public interface ITwoDArray<out T>
+    {
+        T this[int a, int b] { get; }
+    }
+
+    /// <summary>
+    /// Implementation of <see cref="ITwoDArray{T}"/> that acts as a read-only view to an
+    /// actual two-dimensional array.
+    /// </summary>
+    public class TwoDArray<T> : ITwoDArray<T>
+    {
+        private readonly T[,] _array;
+
+        public T this[int a, int b]
+        {
+            get { return _array[a, b]; }
+        }
+
+        public TwoDArray(T[,] array)
+        {
+            _array = array;
+        }
+    }
+
+    /// <summary>
+    /// Implementation of <see cref="ITwoDArray{T}"/> that returns elements generated by a function.
+    /// </summary>
+    public class FunctionalTwoDArray<T> : ITwoDArray<T>
+    {
+        private readonly Func<int, int, T> _f;
+
+        public T this[int a, int b]
+        {
+            get { return _f(a, b); }
+        }
+
+        public FunctionalTwoDArray(Func<int, int, T> f)
+        {
+            _f = f;
+        }
+    }
+}
