+++ WPFSKillTree/TreeGenerator/Algorithm/SteinerReductions/BottleneckSteinerDistanceCalculator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using POESKillTree.TreeGenerator.Algorithm.Model;
 
 namespace POESKillTree.TreeGenerator.Algorithm.SteinerReductions
 {
