+++ WPFSKillTree/TreeGenerator/Algorithm/SteinerReductions/FarAwayNonTerminalsTest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using POESKillTree.TreeGenerator.Algorithm.Model;
 using POESKillTree.Utils;
 
 namespace POESKillTree.TreeGenerator.Algorithm.SteinerReductions
