+++ WPFSKillTree/WPFSKillTree.csproj
@@ -207,9 +207,9 @@
     <Compile Include="SkillTreeFiles\Updater.cs" />
     <Compile Include="TreeGenerator\Algorithm\DisjointSet.cs" />
     <Compile Include="TreeGenerator\Algorithm\DistanceLookup.cs" />
-    <Compile Include="TreeGenerator\Algorithm\GraphEdge.cs" />
-    <Compile Include="TreeGenerator\Algorithm\GraphEdgeSet.cs" />
-    <Compile Include="TreeGenerator\Algorithm\GraphNode.cs" />
+    <Compile Include="TreeGenerator\Algorithm\Model\GraphEdge.cs" />
+    <Compile Include="TreeGenerator\Algorithm\Model\GraphEdgeSet.cs" />
+    <Compile Include="TreeGenerator\Algorithm\Model\GraphNode.cs" />
     <Compile Include="TreeGenerator\Algorithm\HillClimber.cs" />
     <Compile Include="TreeGenerator\Algorithm\LinkedListPriorityQueue.cs" />
     <Compile Include="TreeGenerator\Algorithm\MinimalSpanningTree.cs" />
@@ -218,13 +218,14 @@
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\BottleneckSteinerDistanceCalculator.cs" />
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\DegreeTest.cs" />
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\FarAwayNonTerminalsTest.cs" />
-    <Compile Include="TreeGenerator\Algorithm\Data.cs" />
+    <Compile Include="TreeGenerator\Algorithm\Model\Data.cs" />
     <Compile Include="TreeGenerator\Algorithm\NodeStates.cs" />
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\NearestVertexTest.cs" />
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\NonTerminalsOfDegreeKTest.cs" />
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\PathsWithManyTerminalsTest.cs" />
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\ShortestLinksTest.cs" />
     <Compile Include="TreeGenerator\Algorithm\SteinerReductions\SteinerReduction.cs" />
+    <Compile Include="TreeGenerator\Algorithm\Model\TwoDArray.cs" />
     <Compile Include="TreeGenerator\Algorithm\VoronoiPartition.cs" />
     <Compile Include="TreeGenerator\Genetic\BitArrayKey.cs" />
     <Compile Include="TreeGenerator\Genetic\GeneticAlgorithm.cs" />
